+++ test/HAaveProtocolV2_repay.test.js
@@ -160,7 +160,7 @@ contract('Aave V2', function([_, user]) {
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
       // (borrow - repay) <= debtTokenUserAfter < (borrow + interestMax - repay)
       expect(debtTokenUserAfter).to.be.bignumber.gte(borrowAmount.sub(value));
@@ -206,7 +206,7 @@ contract('Aave V2', function([_, user]) {
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
       // (borrow - repay) <= debtTokenUserAfter < (borrow + interestMax - repay)
       expect(debtTokenUserAfter).to.be.bignumber.gte(borrowAmount.sub(value));
@@ -222,8 +222,8 @@ contract('Aave V2', function([_, user]) {
     });
 
     it('whole', async function() {
-      const value = ether('2');
-      const extraNeed = value.sub(borrowAmount);
+      const extraNeed = ether('1');
+      const value = borrowAmount.add(extraNeed);
       const to = this.hAaveV2.address;
       const data = abi.simpleEncode(
         'repay(address,uint256,uint256,address)',
@@ -255,13 +255,13 @@ contract('Aave V2', function([_, user]) {
       const interestMax = borrowAmount.mul(new BN(1)).div(new BN(10000));
 
       // Verify handler return
-      expect(handlerReturn).to.be.zero;
+      expect(handlerReturn).to.be.bignumber.zero;
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
-      expect(debtTokenUserAfter).to.be.zero;
+      expect(debtTokenUserAfter).to.be.bignumber.zero;
       // (repay - borrow - interestMax) < borrowTokenUserAfter <= (repay - borrow)
       expect(borrowTokenUserAfter).to.be.bignumber.lte(value.sub(borrowAmount));
       expect(borrowTokenUserAfter).to.be.bignumber.gt(
@@ -274,7 +274,8 @@ contract('Aave V2', function([_, user]) {
     });
 
     it('whole by ETH', async function() {
-      const value = ether('2');
+      const extraNeed = ether('1');
+      const value = borrowAmount.add(extraNeed);
       const to = this.hAaveV2.address;
       const data = abi.simpleEncode(
         'repayETH(uint256,uint256,address)',
@@ -297,13 +298,13 @@ contract('Aave V2', function([_, user]) {
       const interestMax = borrowAmount.mul(new BN(1)).div(new BN(10000));
 
       // Verify handler return
-      expect(handlerReturn).to.be.zero;
+      expect(handlerReturn).to.be.bignumber.zero;
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
-      expect(debtTokenUserAfter).to.be.zero;
+      expect(debtTokenUserAfter).to.be.bignumber.zero;
       // (repay - borrow - interestMax) < borrowTokenUserAfter <= (repay - borrow)
       expect(
         (await balanceUser.delta()).add(new BN(receipt.receipt.gasUsed))
@@ -470,7 +471,7 @@ contract('Aave V2', function([_, user]) {
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
       // (borrow - repay) <= debtTokenUserAfter < (borrow + interestMax - repay)
       expect(debtTokenUserAfter).to.be.bignumber.gte(borrowAmount.sub(value));
@@ -519,7 +520,7 @@ contract('Aave V2', function([_, user]) {
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
       // (borrow - repay) <= debtTokenUserAfter < (borrow + interestMax - repay)
       expect(debtTokenUserAfter).to.be.bignumber.gte(borrowAmount.sub(value));
@@ -535,8 +536,8 @@ contract('Aave V2', function([_, user]) {
     });
 
     it('whole', async function() {
-      const value = ether('2');
-      const extraNeed = value.sub(borrowAmount);
+      const extraNeed = ether('1');
+      const value = borrowAmount.add(extraNeed);
       const to = this.hAaveV2.address;
       const data = abi.simpleEncode(
         'repay(address,uint256,uint256,address)',
@@ -568,13 +569,13 @@ contract('Aave V2', function([_, user]) {
       const interestMax = borrowAmount.mul(new BN(1)).div(new BN(10000));
 
       // Verify handler return
-      expect(handlerReturn).to.be.zero;
+      expect(handlerReturn).to.be.bignumber.zero;
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
-      expect(debtTokenUserAfter).to.be.zero;
+      expect(debtTokenUserAfter).to.be.bignumber.zero;
       // (repay - borrow - interestMax) < borrowTokenUserAfter <= (repay - borrow)
       expect(borrowTokenUserAfter).to.be.bignumber.lte(value.sub(borrowAmount));
       expect(borrowTokenUserAfter).to.be.bignumber.gt(
@@ -587,7 +588,8 @@ contract('Aave V2', function([_, user]) {
     });
 
     it('whole by ETH', async function() {
-      const value = ether('2');
+      const extraNeed = ether('1');
+      const value = borrowAmount.add(extraNeed);
       const to = this.hAaveV2.address;
       const data = abi.simpleEncode(
         'repayETH(uint256,uint256,address)',
@@ -610,13 +612,13 @@ contract('Aave V2', function([_, user]) {
       const interestMax = borrowAmount.mul(new BN(1)).div(new BN(10000));
 
       // Verify handler return
-      expect(handlerReturn).to.be.zero;
+      expect(handlerReturn).to.be.bignumber.zero;
       // Verify proxy balance
       expect(
         await this.borrowToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       // Verify user balance
-      expect(debtTokenUserAfter).to.be.zero;
+      expect(debtTokenUserAfter).to.be.bignumber.zero;
       // (repay - borrow - interestMax) < borrowTokenUserAfter <= (repay - borrow)
       expect(
         (await balanceUser.delta()).add(new BN(receipt.receipt.gasUsed))