+++ test/HFurucomboStaking.test.js
@@ -94,9 +94,9 @@ contract('Furucombo', function([_, user, someone]) {
       });
 
       // Verify
-      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.token.balanceOf.call(user)).to.be.zero;
+      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.token.balanceOf.call(user)).to.be.bignumber.zero;
       expect(await balanceUser.delta()).to.be.bignumber.eq(
         ether('0').sub(new BN(receipt.receipt.gasUsed))
       );
@@ -136,9 +136,9 @@ contract('Furucombo', function([_, user, someone]) {
       });
 
       // Verify
-      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await balanceProxy.get()).to.be.zero;
-      expect(await this.token.balanceOf.call(user)).to.be.zero;
+      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
+      expect(await this.token.balanceOf.call(user)).to.be.bignumber.zero;
       expect(await balanceUser.delta()).to.be.bignumber.eq(
         ether('0').sub(new BN(receipt.receipt.gasUsed))
       );
@@ -250,8 +250,8 @@ contract('Furucombo', function([_, user, someone]) {
       const tokenUserAfter = await this.token.balanceOf.call(user);
 
       // Verify
-      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await balanceProxy.get()).to.be.zero;
+      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
       expect(tokenUserAfter.sub(tokenUserBefore)).to.be.bignumber.eq(
         unstakeAmount
       );
@@ -294,12 +294,12 @@ contract('Furucombo', function([_, user, someone]) {
       const tokenUserAfter = await this.token.balanceOf.call(user);
 
       // Verify
-      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
-      expect(await balanceProxy.get()).to.be.zero;
+      expect(await this.token.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
       expect(tokenUserAfter.sub(tokenUserBefore)).to.be.bignumber.eq(
         unstakeAmount
       );
-      expect(await this.staking.balanceOf(user)).to.be.zero;
+      expect(await this.staking.balanceOf(user)).to.be.bignumber.zero;
       expect(await balanceUser.delta()).to.be.bignumber.eq(
         ether('0').sub(new BN(receipt.receipt.gasUsed))
       );
@@ -461,10 +461,10 @@ contract('Furucombo', function([_, user, someone]) {
       expect(rewardUserAfter.sub(rewardUserBefore)).to.be.bignumber.eq(
         claimAmount.add(claimAmount)
       );
-      expect(await balanceProxy.get()).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
       expect(
         await this.rewardToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       expect(await balanceUser.delta()).to.be.bignumber.eq(
         ether('0').sub(new BN(receipt.receipt.gasUsed))
       );
@@ -514,10 +514,10 @@ contract('Furucombo', function([_, user, someone]) {
       expect(rewardUserAfter.sub(rewardUserBefore)).to.be.bignumber.eq(
         claimAmount.add(claimAmount).add(claimAmount)
       );
-      expect(await balanceProxy.get()).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
       expect(
         await this.rewardToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
+      ).to.be.bignumber.zero;
       expect(await balanceUser.delta()).to.be.bignumber.eq(
         ether('0').sub(new BN(receipt.receipt.gasUsed))
       );
@@ -551,11 +551,11 @@ contract('Furucombo', function([_, user, someone]) {
       expect(rewardUserAfter.sub(rewardUserBefore)).to.be.bignumber.eq(
         claimAmount.add(claimAmount)
       );
-      expect(await balanceProxy.get()).to.be.zero;
+      expect(await balanceProxy.get()).to.be.bignumber.zero;
       expect(
         await this.rewardToken.balanceOf.call(this.proxy.address)
-      ).to.be.zero;
-      expect(await balanceUser.delta()).to.be.zero;
+      ).to.be.bignumber.zero;
+      expect(await balanceUser.delta()).to.be.bignumber.zero;
       profileGas(receipt);
     });