+++ test/HPolygon.test.js
@@ -113,7 +113,7 @@ contract('Polygon Token Bridge', function([_, user]) {
         );
         // Verify balance
         expect(await balanceBridge.delta()).to.be.bignumber.eq(value);
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0')
             .sub(value)
@@ -157,7 +157,7 @@ contract('Polygon Token Bridge', function([_, user]) {
         );
         // Verify balance
         expect(await balanceBridge.delta()).to.be.bignumber.eq(value);
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0')
             .sub(value)
@@ -214,10 +214,10 @@ contract('Polygon Token Bridge', function([_, user]) {
           await this.token.balanceOf.call(POLYGON_POS_PREDICATE_ERC20)
         ).to.be.bignumber.eq(tokenBridgeAmount.add(value));
         // Verify Proxy balance
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         // Verify User balance
-        expect(await this.token.balanceOf.call(user)).to.be.zero;
+        expect(await this.token.balanceOf.call(user)).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );
@@ -273,10 +273,10 @@ contract('Polygon Token Bridge', function([_, user]) {
           await this.token.balanceOf.call(POLYGON_POS_PREDICATE_ERC20)
         ).to.be.bignumber.eq(tokenBridgeAmount.add(value));
         // Verify Proxy balance
-        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await this.token.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         // Verify User balance
-        expect(await this.token.balanceOf.call(user)).to.be.zero;
+        expect(await this.token.balanceOf.call(user)).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );
@@ -342,10 +342,10 @@ contract('Polygon Token Bridge', function([_, user]) {
           await this.matic.balanceOf.call(POLYGON_PLASMA_DEPOSIT_MANAGER)
         ).to.be.bignumber.eq(maticBridgeAmount.add(value));
         // Verify Proxy balance
-        expect(await this.matic.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await this.matic.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         // Verify User balance
-        expect(await this.matic.balanceOf.call(user)).to.be.zero;
+        expect(await this.matic.balanceOf.call(user)).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );
@@ -400,10 +400,10 @@ contract('Polygon Token Bridge', function([_, user]) {
           await this.matic.balanceOf.call(POLYGON_PLASMA_DEPOSIT_MANAGER)
         ).to.be.bignumber.eq(maticBridgeAmount.add(value));
         // Verify Proxy balance
-        expect(await this.matic.balanceOf.call(this.proxy.address)).to.be.zero;
-        expect(await balanceProxy.get()).to.be.zero;
+        expect(await this.matic.balanceOf.call(this.proxy.address)).to.be.bignumber.zero;
+        expect(await balanceProxy.get()).to.be.bignumber.zero;
         // Verify User balance
-        expect(await this.matic.balanceOf.call(user)).to.be.zero;
+        expect(await this.matic.balanceOf.call(user)).to.be.bignumber.zero;
         expect(await balanceUser.delta()).to.be.bignumber.eq(
           ether('0').sub(new BN(receipt.receipt.gasUsed))
         );